+++ NEWS.md
@@ -1,3 +1,7 @@
+# Version 0.6.1
+* Fixed a bug for reading files with different overview levels.
+* Added additional early checks for grid shapes.
+
 # Version 0.6.0
 * Added `closed_border` argument to control whether cells outside the mask are excluded from the analysis.
 * Fixed the polygon-masked output to not return the buffered array.